+++ MiniProject/MiniProject.vcxproj
@@ -18,6 +18,15 @@
       <Platform>x64</Platform>
     </ProjectConfiguration>
   </ItemGroup>
+  <ItemGroup>
+    <ClCompile Include="source\handler.cpp" />
+    <ClCompile Include="source\Item.cpp" />
+    <ClCompile Include="source\MiniMain.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="source\handler.h" />
+    <ClInclude Include="source\Item.h" />
+  </ItemGroup>
   <PropertyGroup Label="Globals">
     <VCProjectVersion>16.0</VCProjectVersion>
     <Keyword>Win32Proj</Keyword>
@@ -138,9 +147,6 @@
       <GenerateDebugInformation>true</GenerateDebugInformation>
     </Link>
   </ItemDefinitionGroup>
-  <ItemGroup>
-    <ClCompile Include="source\Case1_for_ClassMember.cpp" />
-  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
