--- MiniProject/source/Case2_for_ClassMember.cpp
@@ -1,54 +1,0 @@
-class Product1
-{
-private:
-	char* CPU_Name;				// CPU 이름
-	char* CPU_Num;				// CPU 품번
-	char* CPU_Searial;			// CPU 시리얼번호
-
-	char* MB_Name;				// 메인보드 이름
-	char* MB_Num;				// 메인보드 품번
-	char* MB_Searial;			// 메인보드 시리얼번호
-
-	char* Memory_Name;			// 메모리 이름
-	char* Memory_Num;			// 메모리 품번
-	char* Memory_Searial;		// 메모리 시리얼번호
-
-	char* HDD_Name;				// HDD 이름
-	char* HDD_Num;				// HDD 품번
-	char* HDD_Searial;			// HDD 시리얼번호
-
-	char* SSD_Name;				// SSD 이름
-	char* SSD_Num;				// SSD 품번
-	char* SSD_Searial;			// SSD 시리얼번호
-
-	char* Case_Name;			// 케이스 이름
-	char* Case_Num;				// 케이스 품번
-	char* Case_Searial;			// 케이스 시리얼번호
-
-	char* Power_Name;			// 파워 이름
-	char* Power_Num;			// 파워 품번
-	char* Power_Searial;		// 파워 시리얼번호
-
-	char* Software_Name;		// 소프트웨어(운영체제) 이름
-	char* Software_Num;			// 소프트웨어(운영체제) 품번
-	char* Software_Searial;		// 소프트웨어(운영체제) 시리얼번호
-
-public:
-	Product1()
-	{
-
-	}
-
-	void Print_ProductInfo() const
-	{
-
-	}
-
-	void Get_CPUInfo() const
-	{
-
-	}
-
-
-
-};
